requirements.txt
# Web scraping and HTTP requests
requests
playwright
beautifulsoup4

# PDF processing
PyPDF2
pdfplumber

# Configuration and logging
python-dotenv
pyyaml

# Data handling
pandas
openpyxl
sqlalchemy

# Geographic processing
geopy

# AI/ML integration (optional for self-learning)
scikit-learn

# API clients (for future AI agent integration)
httpx
anthropic
openai

# CLI and interface libraries
rich
click
tqdm
tabulate

# Additional utilities for hybrid system and portal authentication
cryptography  # For secure credential storage